Update the main JS file
