--- book.js
@@ -1,7 +1,0 @@
-// book class that represents the book
-export default class Book {
-  constructor(title, author) {
-    this.title = title;
-    this.author = author;
-  }
-}
